Removed some code repetition
