+++ src/main/java/dev/tycho/stonks/managers/DatabaseManager.java
@@ -9,6 +9,7 @@ import com.j256.ormlite.table.TableUtils;
 import dev.tycho.stonks.Stonks;
 import dev.tycho.stonks.command.CompanyCommand;
 import dev.tycho.stonks.database.*;
+import dev.tycho.stonks.model.accountvisitors.IAccountVisitor;
 import dev.tycho.stonks.model.logging.Transaction;
 import dev.tycho.stonks.model.core.*;
 import dev.tycho.stonks.model.service.Service;
@@ -153,6 +154,31 @@ public class DatabaseManager extends SpigotModule {
       e.printStackTrace();
       Bukkit.broadcastMessage(ChatColor.RED + "SQL Exception creating a log ");
     }
+  }
+  public void updateAccount(Account account) {
+    //Update the account database
+    IAccountVisitor visitor = new IAccountVisitor() {
+      @Override
+      public void visit(CompanyAccount a) {
+        try {
+          getCompanyAccountDao().update(a);
+        } catch (SQLException e) {
+          e.printStackTrace();
+        }
+      }
+
+      @Override
+      public void visit(HoldingsAccount a) {
+        try {
+          //Update the account and holdings
+          getHoldingAccountDao().update(a);
+          for (Holding h : a.getHoldings()) getHoldingDao().update(h);
+        } catch (SQLException e) {
+          e.printStackTrace();
+        }
+      }
+    };
+    account.accept(visitor);
   }
 
   public Dao<Service, Integer> getServiceDao() {
